requirements/base.txt
fastapi>=0.109.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
rembg[cpu]>=2.0.53
onnxruntime>=1.17.0
pillow>=10.2.0
numpy>=1.26.0
redis>=5.0.1
rq>=1.15.0
requests>=2.31.0
